move bias_to_trigger_patch_map tp pixels; remove pandas dep; use lru_cache
